extracted Superblock struct into separate file
